Un-OO storage
Instead of implementing the low level storage as class, refactor the
storage creation into an function that returns an simple object with
methods for the necessary features. This hides the "private" methods
that JavaScript really doesn't support and makes the API more compatible
with varasto-server package that uses same kind of function for
constructing the Express.js stuff.