Introduce namespaces
Replace current implementation of the low level storage with a new one
that uses namespaces.

Each item stored into the storage must have an namespace and key that
must be valid URL slugs. Other than that, the functionality remains the
same, although the methods have been renamed to match ones in the Map
class.